+++ src/scandeval/benchmarker.py
@@ -48,7 +48,7 @@ class BenchmarkConfigParams(BaseModel):
     verbose: bool
     trust_remote_code: bool
     load_in_4bit: bool | None
-    use_flash_attention: bool
+    use_flash_attention: bool | None
     clear_model_cache: bool
     only_validation_split: bool
     few_shot: bool
@@ -155,7 +155,7 @@ class Benchmarker:
         verbose: bool = False,
         trust_remote_code: bool = False,
         load_in_4bit: bool | None = None,
-        use_flash_attention: bool = False,
+        use_flash_attention: bool | None = None,
         clear_model_cache: bool = False,
         only_validation_split: bool = False,
         few_shot: bool = True,
@@ -225,7 +225,8 @@ class Benchmarker:
                 done if CUDA is available and the model is a decoder model. Defaults to
                 None.
             use_flash_attention:
-                Whether to use Flash Attention. Defaults to False.
+                Whether to use Flash Attention. If None then it will be used if it is
+                installed and the model is a decoder model. Defaults to None.
             clear_model_cache:
                 Whether to clear the model cache after benchmarking each model.
                 Defaults to False.
@@ -279,7 +280,7 @@ class Benchmarker:
         )
 
         self.benchmark_config = build_benchmark_config(
-            **self.benchmark_config_default_params.model_dump()
+            first_time=True, **self.benchmark_config_default_params.model_dump()
         )
 
         # Initialise variable storing model lists, so we only have to fetch it once
@@ -568,7 +569,7 @@ class Benchmarker:
         self,
         model: list[str] | str | None,
         model_languages: list[Language],
-        token: bool | str,
+        token: bool | str | None,
     ) -> list[str]:
         """Prepare the model ID(s) to be benchmarked.
 
@@ -687,7 +688,9 @@ class Benchmarker:
         """Call the benchmarker. See `Benchmarker.benchmark`."""
         return self.benchmark(*args, **kwargs)
 
-    def _get_model_ids(self, languages: list[Language], token: bool | str) -> list[str]:
+    def _get_model_ids(
+        self, languages: list[Language], token: bool | str | None
+    ) -> list[str]:
         """Get list of model IDs from the Hugging Face Hub.
 
         Args: